feat: cleanup code
